added back-end: -signing up -login -logout -passport authentication -logged in condition for to do list -repathing back to login

+++ controllers/tasksController.js
@@ -1,7 +1,10 @@
+
 const pool = require('../models/database');
+const { v4: uuidv4 } = require('uuid');
 
 exports.getToDoTasks = (req, res) => {
-    pool.query('SELECT * FROM tasks WHERE status = ($1)', ['todo'], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('SELECT * FROM tasks WHERE status = ($1) AND user_id = $2', ['todo', user_id], (error, results) => {
         if (error) {
             throw error;
         }
@@ -10,7 +13,8 @@ exports.getToDoTasks = (req, res) => {
 };
 
 exports.getInProgressTasks = (req, res) => {
-    pool.query('SELECT * FROM tasks WHERE status = ($1)', ['inprogress'], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('SELECT * FROM tasks WHERE status = ($1) AND user_id = $2', ['inprogress', user_id], (error, results) => {
         if (error) {
             throw error;
         }
@@ -19,7 +23,8 @@ exports.getInProgressTasks = (req, res) => {
 }; 
 
 exports.getFinishedTasks = (req, res) => {
-    pool.query('SELECT * FROM tasks WHERE status = ($1)', ['finished'], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('SELECT * FROM tasks WHERE status = ($1) AND user_id = $2', ['finished', user_id], (error, results) => {
         if (error) {
             throw error;
         }
@@ -27,19 +32,10 @@ exports.getFinishedTasks = (req, res) => {
     });
 };
 
-exports.createTask = (req, res) => {
-    const { id, name, description, status } = req.body;
-    pool.query('INSERT INTO tasks (id, name, description, status) VALUES ($1, $2, $3, $4)', [id, name, description, status], (error, results) => {
-        if (error) {
-            throw error;
-        }
-        res.status(201).send(`Task added`);
-    });
-};
-
 exports.deleteTask = (req, res) => {
     const id = req.params.id;
-    pool.query('DELETE FROM tasks WHERE id = $1', [id], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('DELETE FROM tasks WHERE id = $1 AND user_id = $2', [id, user_id], (error, results) => {
         if (error) {
             throw error;
         }
@@ -49,7 +45,8 @@ exports.deleteTask = (req, res) => {
 
 exports.updateTaskToToDo = (req, res) => {
     const id = req.params.id;
-    pool.query('UPDATE tasks SET status = $1 WHERE id = $2', ['todo', id], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('UPDATE tasks SET status = $1 WHERE id = $2 AND user_id = $3', ['todo', id, user_id], (error, results) => {
         if (error) {
             throw error;
         }
@@ -59,7 +56,8 @@ exports.updateTaskToToDo = (req, res) => {
 
 exports.updateTaskToInProgress = (req, res) => {
     const id = req.params.id;
-    pool.query('UPDATE tasks SET status = $1 WHERE id = $2', ['inprogress', id], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('UPDATE tasks SET status = $1 WHERE id = $2 AND user_id = $3', ['inprogress', id, user_id], (error, results) => {
         if (error) {
             throw error;
         }
@@ -69,10 +67,55 @@ exports.updateTaskToInProgress = (req, res) => {
 
 exports.updateTaskToFinished = (req, res) => {
     const id = req.params.id;
-    pool.query('UPDATE tasks SET status = $1 WHERE id = $2', ['finished', id], (error, results) => {
+    const user_id = req.user.id;
+    pool.query('UPDATE tasks SET status = $1 WHERE id = $2 AND user_id = $3', ['finished', id, user_id], (error, results) => {
         if (error) {
             throw error;
         }
         res.status(200).send(`Status of task with ID ${id} updated to finished`);
+    });
+};
+
+exports.createTask = (req, res) => {
+    const { id, name, description, status } = req.body;
+    const user_id = req.user.id;
+    pool.query('INSERT INTO tasks (id, name, description, status, user_id) VALUES ($1, $2, $3, $4, $5)', [id, name, description, status, user_id], (error, results) => {
+        if (error) {
+            throw error;
+        }
+        res.status(201).send(`Task added`);
+    });
+};
+
+exports.addDefaultTasks = (req, res) => {
+    const { user_id } = req.body;
+    const defaultTasks = [
+        {
+            id: uuidv4(),
+            name: 'this is a to do task',
+            description: 'The buttons to the right allow you to:\n1) delete this task\n2) move this task to the in progress list\n3) move this task to the finished list',
+            status: 'todo',
+            user_id
+        },
+        {
+            id: uuidv4(),
+            name: 'this is an in progress task',
+            description: 'The buttons to the right allow you to:\n1) delete this task\n2) move this task to the to do list\n3) move this task to the finished list',
+            status: 'inprogress',
+            user_id
+        },
+        {
+            id: uuidv4(),
+            name: 'this is a finished task',
+            description: 'The buttons to the right allow you to:\n1) delete this task\n2) move this task to the to do list\n3) move this task to the in progress list',
+            status: 'finished',
+            user_id
+        },
+    ];
+    pool.query('INSERT INTO tasks (id, name, description, status, user_id) VALUES ($1, $2, $3, $4, $5), ($6, $7, $8, $9, $10), ($11, $12, $13, $14, $15)', [defaultTasks[0].id, defaultTasks[0].name, defaultTasks[0].description, defaultTasks[0].status, defaultTasks[0].user_id, defaultTasks[1].id, defaultTasks[1].name, defaultTasks[1].description, defaultTasks[1].status, defaultTasks[1].user_id, defaultTasks[2].id, defaultTasks[2].name, defaultTasks[2].description, defaultTasks[2].status, defaultTasks[2].user_id], (error, results) => {
+        if (error) {
+            throw error;
+        }
+        res.status(201).send(`Default tasks added`);
     });
 };
